lib: generic: zynqmp_a78: Add support for versal_net a78 platform with System Device Tree workflow
System Device Tree workflow is AMD-Xilinx workflow whereby the BSP, libraries
and applications in software are derived from a hardware-design. The
hardware-design is used to generate a system device tree (SDT) that
describes information for Linux and other processing environments.

Additionally:
- the xreg/xcpu files are not generated for SDT workflow. Because of
this do not include if SDT symbol is present in BSP.
- As MB and GB are provided by lib/utilities.h instead reference from there.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>

tanmay.shah@amd.com