examples: freertos: define xInterruptController in System Device Tree Flow
When building with freertos demo in System Device Tree (SDT) workflow, the
application fails to build because it is set to extern. This symbol is not
found in the SDT BSP. Match generic Libmetal AMP Demo and set this symbol
to static to fix compilation issue.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>

tanmay.shah@amd.com